Housekeeping. Do some renaming, remove some asset files from Git tracking.

+++ gcrt/src/rtscene.cpp
@@ -61,7 +61,7 @@ void RTScene::Add(RTMaterial mat)
 {
     assert(matIdcs.count(mat.name) == 0);
     mats.push_back(mat);
-    matIdcs[mat.name] = mats.size() - 1;
+    matIdcs[mat.name] = (uint32_t)mats.size() - 1;
 }
 
 /**
